Change RenderDraw::m_streamMask to uint32_t from uint8_t to allow for maximum 32 streams

+++ src/bgfx.cpp
@@ -1229,7 +1229,7 @@ namespace bgfx
 		m_draw.m_uniformBegin = m_uniformBegin;
 		m_draw.m_uniformEnd   = m_uniformEnd;
 
-		if (UINT8_MAX != m_draw.m_streamMask)
+		if (UINT32_MAX != m_draw.m_streamMask)
 		{
 			uint32_t numVertices = UINT32_MAX;
 			for (uint32_t idx = 0, streamMask = m_draw.m_streamMask
